succesfully extracted patch API request to api interface file, have left notes
